pyproject.toml
[project]
name = "langchainsandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.15.0",
    "ipython>=8.32.0",
    "langchain>=0.3.18",
    "langchain-anthropic>=0.3.7",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.34",
    "langchain-google-genai>=2.0.9",
    "langchain-localai>=0.1.2",
    "langchain-mistralai>=0.2.6",
    "langchain-openai>=0.3.4",
    "langgraph>=0.2.70",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
]
